Refactor STL data file processing
